backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.14",
    "pytest>=8.4.1",
    "uvicorn>=0.34.3",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "bcrypt==3.2.2",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "python-jose>=3.3.0",
    "email-validator>=2.0.0",
    "python-multipart>=0.0.5",
    "alembic>=1.16.2",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy-utils>=0.41.2",
    "pytest-asyncio>=1.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
